Query the GraphQL engine directly
