+++ src/test/kotlin/JwtTest.kt
@@ -8,19 +8,22 @@ import java.sql.Timestamp
 import java.time.LocalDateTime
 
 class JwtTest : FunSpec({
-    /** Tests that the [actual] [LocalDateTime] is within five seconds of the [expected] [LocalDateTime]. */
-    fun testDateTime(actual: LocalDateTime, expected: LocalDateTime) {
-        val leewayInSeconds = 5L
-        actual shouldBeAfter expected.minusSeconds(leewayInSeconds)
-        actual shouldBeBefore expected.plusSeconds(leewayInSeconds)
-    }
+    context("build(String, LocalDateTime)") {
+        /** Tests that the [actual] [LocalDateTime] is within five seconds of the [expected] [LocalDateTime]. */
+        fun testDateTime(actual: LocalDateTime, expected: LocalDateTime) {
+            val leewayInSeconds = 5L
+            actual shouldBeAfter expected.minusSeconds(leewayInSeconds)
+            actual shouldBeBefore expected.plusSeconds(leewayInSeconds)
+        }
 
-    fun readExpiry(token: String): LocalDateTime = JWT.decode(token).expiresAt.run { Timestamp(time).toLocalDateTime() }
+        fun readExpiry(token: String): LocalDateTime =
+            JWT.decode(token).expiresAt.run { Timestamp(time).toLocalDateTime() }
 
-    test("The access and refresh tokens should expire in one hour and one week respectively") {
-        val (accessToken, refreshToken) = buildAuthToken("user ID")
-        val now = LocalDateTime.now()
-        testDateTime(actual = readExpiry(accessToken), expected = now.plusHours(1))
-        testDateTime(actual = readExpiry(refreshToken), expected = now.plusWeeks(1))
+        test("The access and refresh tokens should expire in one hour and one week respectively") {
+            val (accessToken, refreshToken) = buildAuthToken("user ID")
+            val now = LocalDateTime.now()
+            testDateTime(actual = readExpiry(accessToken), expected = now.plusHours(1))
+            testDateTime(actual = readExpiry(refreshToken), expected = now.plusWeeks(1))
+        }
     }
 })
