+++ src/test/kotlin/ModelsUtil.kt
@@ -1,7 +1,5 @@
 package com.neelkamath.omniChat
 
-fun Account.toUpdatedAccount(): UpdatedAccount = UpdatedAccount(id, username, emailAddress, firstName, lastName)
-
 fun buildNewGroupChat(userIdList: List<String>): NewGroupChat =
     NewGroupChat(GroupChatTitle("T"), GroupChatDescription(""), userIdList.toList())
