+++ src/test/kotlin/graphql/routing/SubscriptionsTest.kt
@@ -5,10 +5,9 @@ import com.neelkamath.omniChat.GraphQlRequest
 import com.neelkamath.omniChat.createVerifiedUsers
 import com.neelkamath.omniChat.db.tables.Contacts
 import com.neelkamath.omniChat.graphql.operations.CREATED_SUBSCRIPTION_FRAGMENT
-import com.neelkamath.omniChat.graphql.operations.subscriptions.SUBSCRIBE_TO_CONTACTS_QUERY
-import com.neelkamath.omniChat.graphql.operations.subscriptions.operateGraphQlSubscription
-import com.neelkamath.omniChat.graphql.operations.subscriptions.parseFrameData
-import com.neelkamath.omniChat.graphql.operations.subscriptions.subscribeToContacts
+import com.neelkamath.omniChat.graphql.operations.DELETED_CONTACT_FRAGMENT
+import com.neelkamath.omniChat.graphql.operations.NEW_CONTACT_FRAGMENT
+import com.neelkamath.omniChat.graphql.operations.UPDATED_CONTACT_FRAGMENT
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.nulls.shouldBeNull
 import io.kotest.matchers.nulls.shouldNotBeNull
@@ -33,7 +32,7 @@ class SubscriptionsTest : FunSpec({
             """
             val operationName = if (shouldSupplyOperationName) "SubscribeToContacts" else null
             val token = createVerifiedUsers(1)[0].accessToken
-            operateGraphQlSubscription(
+            executeGraphQlSubscriptionViaWebSocket(
                 uri = "contacts-subscription",
                 request = GraphQlRequest(query, operationName = operationName),
                 accessToken = token
@@ -60,19 +59,23 @@ class SubscriptionsTest : FunSpec({
         ) { testOperationName(shouldSupplyOperationName = false) }
     }
 
-    context("closeWithError(DefaultWebSocketServerSession, ExecutionResult)") {
-        test(
-            """
-            Given an operation requiring an access token,
-            when calling the operation without a token,
-            then the connection should be closed with the status code 1008
-            """
-        ) {
-            operateGraphQlSubscription(
-                uri = "contacts-subscription",
-                request = GraphQlRequest(SUBSCRIBE_TO_CONTACTS_QUERY)
-            ) { incoming -> incoming.receive().frameType shouldBe FrameType.CLOSE }
-        }
+    fun subscribeToContacts(accessToken: String? = null, callback: SubscriptionCallback) {
+        val subscribeToContactsQuery = """
+            subscription SubscribeToContacts {
+                subscribeToContacts {
+                    $CREATED_SUBSCRIPTION_FRAGMENT
+                    $NEW_CONTACT_FRAGMENT
+                    $UPDATED_CONTACT_FRAGMENT
+                    $DELETED_CONTACT_FRAGMENT
+                }
+            }
+        """
+        executeGraphQlSubscriptionViaWebSocket(
+            uri = "contacts-subscription",
+            request = GraphQlRequest(subscribeToContactsQuery),
+            accessToken = accessToken,
+            callback = callback
+        )
     }
 
     context("subscribe(DefaultWebSocketServerSession, GraphQlSubscription, ExecutionResult)") {
@@ -90,6 +93,18 @@ class SubscriptionsTest : FunSpec({
                 incoming.poll().shouldNotBeNull()
                 incoming.poll().shouldBeNull()
             }
+        }
+    }
+
+    context("closeWithError(DefaultWebSocketServerSession, ExecutionResult)") {
+        test(
+            """
+            Given an operation requiring an access token,
+            when calling the operation without a token,
+            then the connection should be closed
+            """
+        ) {
+            subscribeToContacts { incoming -> incoming.receive().frameType shouldBe FrameType.CLOSE }
         }
     }
 })
